finally get the code right to handle array length offset information

+++ libseed/seed-types.h
@@ -51,7 +51,7 @@ JSValueRef seed_value_from_gi_argument_full (JSContextRef ctx,
 				        GArgument * arg,
 				        GITypeInfo * type_info,
 				        JSValueRef * exception,
-                                        gint array_len,
+                                        guint64 array_len,
                                         GITypeTag gi_tag);
 gboolean seed_gi_release_arg (GITransfer transfer,
 			      GITypeInfo * type_info, GArgument * arg);
